test_project/requirements.txt
# Django and libs
Django==4.1.7
djangorestframework==3.14.0
django-environ==0.10.0
django-choices==1.7.2
Pillow==9.4.0
celery==5.3.0
redis==3.5.3
django-cors-headers==4.1.0

# DataBase
psycopg2-binary==2.9.5


# Registration and authentication
djangorestframework-simplejwt==5.2.2
djoser==2.2.0